applied kage to comments; stubbed out email verification

+++ jaga/php/view/comment.view.php
@@ -15,34 +15,37 @@ class CommentView {
 			$username = urlencode($user->username);
 			$userDisplayName = $user->getUserDisplayName();
 			
-			$html .= "\n\t<!-- START COMMENT -->\n";
-			$html .= "\t<div class=\"container\">\n\n";
-				$html .= "\t\t<div class=\"panel panel-default\">\n";
-					
-					$html .= "\t\t\t<div class=\"panel-heading jagaCommentPanelHeading\">\n";
-					
-						$html.= "\t\t\t\t<div class=\"row\">\n";
-						
-							$html .= "\t\t\t\t\t<div class=\"col-md-6\">";
-								$html .= "<b><a href=\"/u/" . $username . "/\">" . $userDisplayName . "</a></b> (<i>" . $comment->commentDateTime .= "</i>)";
-							$html .= "</div>\n";
-							
-							$html .= "\t\t\t\t\t<div class=\"col-md-6\">";						
-								if ($userID == $_SESSION['userID'] || Authentication::isAdmin()) {
-									$html .= "<a href=\"/k/comment/delete/" . $commentID . "/\" class=\"btn btn-default btn-xs pull-right\"><span class=\"glyphicon glyphicon-remove\" style=\"color:#f00;\"></span> DELETE</a>";
-								}
-							$html .= "</div>\n";
-							
-						$html .= "\t\t\t\t</div>\n";
-						
-					$html .= "\t\t\t</div>\n";
-					
-					$html .= "\t\t\t<div class=\"panel-body\">" . nl2br($comment->commentContent) . "</div>\n";
-					
-				$html .= "\t\t</div>\n";
-			$html .= "\n\t</div>\n";
-			$html .= "\t<!-- END COMMENT -->\n\n";
+			if (!$user->userShadowBan || $userID == $_SESSION['userID']) {
 			
+				$html .= "\n\t<!-- START COMMENT -->\n";
+				$html .= "\t<div class=\"container\">\n\n";
+					$html .= "\t\t<div class=\"panel panel-default\">\n";
+						
+						$html .= "\t\t\t<div class=\"panel-heading jagaCommentPanelHeading\">\n";
+						
+							$html.= "\t\t\t\t<div class=\"row\">\n";
+							
+								$html .= "\t\t\t\t\t<div class=\"col-md-6\">";
+									$html .= "<b><a href=\"/u/" . $username . "/\">" . $userDisplayName . "</a></b> (<i>" . $comment->commentDateTime .= "</i>)";
+								$html .= "</div>\n";
+								
+								$html .= "\t\t\t\t\t<div class=\"col-md-6\">";						
+									if ($userID == $_SESSION['userID'] || Authentication::isAdmin()) {
+										$html .= "<a href=\"/k/comment/delete/" . $commentID . "/\" class=\"btn btn-default btn-xs pull-right\"><span class=\"glyphicon glyphicon-remove\" style=\"color:#f00;\"></span> DELETE</a>";
+									}
+								$html .= "</div>\n";
+								
+							$html .= "\t\t\t\t</div>\n";
+							
+						$html .= "\t\t\t</div>\n";
+						
+						$html .= "\t\t\t<div class=\"panel-body\">" . nl2br($comment->commentContent) . "</div>\n";
+						
+					$html .= "\t\t</div>\n";
+				$html .= "\n\t</div>\n";
+				$html .= "\t<!-- END COMMENT -->\n\n";
+				
+			}
 		}
 		
 		return $html;
@@ -88,6 +91,8 @@ class CommentView {
 								$html .= "</div>\n";
 							$html .= "\t\t\t\t\t\t</div>\n";
 
+							$html .= '<div id="hachimitsu"><input type="text" name="comment" autocomplete="off"><input type="url" name="url" autocomplete="off"></div>';
+							
 							$html .= "\t\t\t\t\t\t<div class=\"row\">\n";
 								$html .= "\t\t\t\t\t\t\t<div class=\"col-sm-12\" style=\"margin-top:5px;\">\n";
 									$html .= "\t\t\t\t\t\t\t\t<input type=\"submit\" name=\"jagaContentCommentSubmit\" id=\"jagaContentCommentSubmit\" class=\"btn btn-default jagaFormButton col-xs-8 col-sm-6 col-md-4 pull-right\" value=\"" . Lang::getLang('submitComment') . "\">\n";
